% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcompilation-package.R
\docType{package}
\name{modelcompilation-package}
\alias{modelcompilation-package}
\alias{modelcompilation}
\title{The 'modelcompilation' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
